+++ libraries/MAX31850/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Change Log MAX44009
+# Change Log MAX31850
 
 All notable changes to this project will be documented in this file.
 
@@ -6,12 +6,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.2] - 2023-01-23
+- update GitHub actions
+- update license 2023
+- add setTypeTC() + getTypeTC()
+- update keywords.txt
+- update readme.md
+- add retries parameter to **begin()**
+- minor edits.
+
+
 ## [0.1.1] - 2022-11-16
 - add RP2040 in build-CI
 - add changelog.md
 - major rewrite based on reread of datasheet.
 - initial release(?)
-
 
 ## [0.1.0] - 2021-06-03
 - initial version (not released)